# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Booking API",
        "description": "API Documentation for Event Booking System",
        "contact": {
            "email": "amansrivastava8355@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/attendees": {
            "get": {
                "summary": "List all attendees",
                "operationId": "a0f4e1a2b347e7fa54531a47d05ab3ee",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EventResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Register attendee",
                "operationId": "caa45abe11a508c27e8de3f2c80cd64a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAttendeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttendeeResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create booking",
                "operationId": "893be3fcfe94b319a05e6e01eed19c8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/events": {
            "get": {
                "summary": "List all events",
                "operationId": "1031372a8175e66ba1b46a820fe040d0",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EventResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new event",
                "operationId": "ac2effdd73571c2df8cbc83f879d318e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/events/{id}": {
            "get": {
                "summary": "Get event details",
                "operationId": "f6e17a4fe670cb397894b247c6c5d75f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update event",
                "operationId": "101a7775adde6d5a249038616701d479",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete event",
                "operationId": "7921b70f326e3bc012507123523e4ee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "StoreAttendeeRequest": {
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    }
                },
                "type": "object"
            },
            "AttendeeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "registered_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreBookingRequest": {
                "required": [
                    "event_id",
                    "attendee_id"
                ],
                "properties": {
                    "event_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "attendee_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "notes": {
                        "type": "string",
                        "example": "Front row seat requested"
                    }
                },
                "type": "object"
            },
            "BookingResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "event_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "attendee_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "notes": {
                        "type": "string",
                        "example": "Front row seat requested"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "event": {
                        "$ref": "#/components/schemas/EventResource"
                    },
                    "attendee": {
                        "$ref": "#/components/schemas/AttendeeResource"
                    }
                },
                "type": "object"
            },
            "EventResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Tech Meetup"
                    },
                    "description": {
                        "type": "string",
                        "example": "An event for tech enthusiasts"
                    },
                    "location": {
                        "type": "string",
                        "example": "US"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T10:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T12:00:00Z"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 100
                    },
                    "available_seats": {
                        "type": "integer",
                        "example": 85
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreEventRequest": {
                "required": [
                    "title",
                    "description",
                    "location",
                    "start_time",
                    "end_time",
                    "capacity"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Tech Meetup"
                    },
                    "description": {
                        "type": "string",
                        "example": "An event for tech enthusiasts"
                    },
                    "location": {
                        "type": "string",
                        "example": "US"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T10:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T12:00:00Z"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "UpdateEventRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Updated Title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description"
                    },
                    "location": {
                        "type": "string",
                        "example": "Canada"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-20T10:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-20T12:00:00Z"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Bookings"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ]
}